Updated methods and added UnitTest

+++ BusinessLogicLayer/Services/ProjectService.cs
@@ -1,15 +1,12 @@
 ﻿using AutoMapper;
 using BusinessLogicLayer.Entities;
 using BusinessLogicLayer.Interfaces;
-using BusinessLogicLayer.Models;
 using DataAccessLayer.EF;
 using DataAccessLayer.Interfaces;
 using DataAccessLayer.Models;
-using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace BusinessLogicLayer.Services
@@ -19,19 +16,21 @@ namespace BusinessLogicLayer.Services
         private readonly IProjectRepository _projectRepository;
         private readonly IProjectTaskRepository _projectTaskRepository; 
         private readonly IMapper _mapper;
-        private readonly TaskTrackerDbContext _context;
 
-        public ProjectService(IProjectRepository projectRepository, IProjectTaskRepository projectTaskRepository, IMapper mapper, TaskTrackerDbContext context)
+        public ProjectService(IProjectRepository projectRepository, IProjectTaskRepository projectTaskRepository, IMapper mapper)
         {
             _projectRepository = projectRepository;
             _projectTaskRepository = projectTaskRepository;
             _mapper = mapper;
-            _context = context;
         }
 
         public async Task<Project>AddProjectAsync(Project project)
         {
+            if (project == null)
+                throw new Exception("Project not found!");
+
             var newProject = _mapper.Map<ProjectDto>(project);
+
             await _projectRepository.AddProjectAsync(newProject);
             
             return _mapper.Map<Project>(newProject);
@@ -40,7 +39,11 @@ namespace BusinessLogicLayer.Services
         public async Task<Project> GetProjectAsync(int projectId)
         {
             var expectedProject = await _projectRepository.GetProjectAsync(projectId);
-            return _mapper.Map<Project>(expectedProject);
+
+            if(expectedProject != null)
+                return _mapper.Map<Project>(expectedProject);
+            else
+                throw new Exception($"Project with Id: {projectId} not found!");
         }
 
         public async Task<IEnumerable<Project>> GetAllProjectsAsync()
@@ -57,6 +60,9 @@ namespace BusinessLogicLayer.Services
 
         public async Task UpdateProjectAsync(int projectId, Project project)
         {
+            if (project == null)
+                throw new Exception("Project not found!");
+
             var newProject = _mapper.Map<ProjectDto>(project);
             
             await _projectRepository.UpdateProjectAsync(projectId, newProject);
@@ -64,58 +70,19 @@ namespace BusinessLogicLayer.Services
 
         public async Task DeleteProjectAsync(int projectId)
         {
+            var project = await GetProjectAsync(projectId);
+            if (project == null)
+                throw new Exception("Project not found!");
+            
             await _projectRepository.DeleteProjectAsync(projectId);
         }
 
         public IEnumerable<Project> GetFilter(ProjectFilters search = null)
         {
-            var entity = _context.Projects.AsQueryable();
+            var result = _projectRepository.GetFilter(search);
 
-            if (!string.IsNullOrWhiteSpace(search?.Name))
-            {
-                entity = entity.Where(x => x.Name.Contains(search.Name));
-            }
-
-            if (search.StartDate.HasValue)
-            {
-                entity = entity.Where(x => x.StartDate.Value == search.StartDate.Value);
-            }
-
-            if (search.CompletionDate.HasValue)
-            {
-                entity = entity.Where(x => x.CompletionDate == search.StartDate);
-            }
-
-            if (search.Priority.HasValue)
-            {
-                entity = entity.Where(x => x.Priority >= search.Priority);
-            }
-
-
-            var filterList = new List<Project>();
-
-            foreach (var project in entity)
-            {
-                filterList.Add(_mapper.Map<Project>(project));
-            }
-            return filterList;
+            return _mapper.Map<IEnumerable<Project>>(result);
         }
 
-        public async Task DeleteProjectTaskAsync(int projectTaskId)
-        {
-            await _projectTaskRepository.DeleteProjectTaskAsync(projectTaskId);
-        }
-
-        public async Task<ProjectTaskDto> GetProjectTaskAsync(int projectTaskId)
-        {
-            return await _context.ProjectTasks.FindAsync(projectTaskId);
-        }
-        public async Task<ProjectTaskDto> AddProjectTaskAsync(ProjectTaskDto projectTask)
-        {
-            await _context.ProjectTasks.AddAsync(projectTask);
-
-            await _context.SaveChangesAsync();
-            return projectTask;
-        }
     }
 }
